Rename the files.
